removed item from character inventory
The end game item, "teleportation remote" was accidentally left in the player's inventory. The player should not be able to obtain this, "teleportation remote" until the end of the game, but it was in the player's inventory at the start of the game, so it had to be removed from the player's inventory.